Disable migrations

+++ Migrations/Postgresql/Version20230117134142.php
@@ -18,6 +18,11 @@ final class Version20230117134142 extends AbstractMigration
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'postgresql', 'Migration can only be executed safely on "postgresql".');
 
+        $this->skipIf(
+            !\array_key_exists('neos_contentrepository_domain_model_workspace', $this->sm->listTableNames()),
+            'These migrations are not supported in Neos 9'
+        );
+
         // Remove all disconnected workspace details entities or the foreign table constraint will fail
         $this->addSql('DELETE FROM shel_neos_workspacemodule_domain_model_workspacedetails WHERE workspacename NOT IN (SELECT name FROM neos_contentrepository_domain_model_workspace)');
 
@@ -35,9 +40,15 @@ final class Version20230117134142 extends AbstractMigration
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'postgresql', 'Migration can only be executed safely on "postgresql".');
 
-        $this->addSql('DROP TABLE shel_neos_workspacemodule_domain_model_workspace_1536f_acl_join');
-        $this->addSql('ALTER TABLE shel_neos_workspacemodule_domain_model_workspacedetails DROP CONSTRAINT FK_923CCEA8D940019');
-        $this->addSql('DROP INDEX flow_identity_shel_neos_workspacemodule_domain_model_work_e2fe7');
-        $this->addSql('ALTER TABLE shel_neos_workspacemodule_domain_model_workspacedetails RENAME COLUMN workspace TO workspacename');
+        $tables = $this->sm->listTableNames();
+        if (\array_key_exists('shel_neos_workspacemodule_domain_model_workspace_1536f_acl_join', $tables)) {
+            $this->addSql('DROP TABLE shel_neos_workspacemodule_domain_model_workspace_1536f_acl_join');
+        }
+
+        if (\array_key_exists('shel_neos_workspacemodule_domain_model_workspacedetails', $tables)) {
+            $this->addSql('ALTER TABLE shel_neos_workspacemodule_domain_model_workspacedetails DROP CONSTRAINT FK_923CCEA8D940019');
+            $this->addSql('DROP INDEX flow_identity_shel_neos_workspacemodule_domain_model_work_e2fe7');
+            $this->addSql('ALTER TABLE shel_neos_workspacemodule_domain_model_workspacedetails RENAME COLUMN workspace TO workspacename');
+        }
     }
 }
